backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
scapy>=2.5.0
python-nmap>=0.7.1
manuf>=1.1.5
icmplib>=3.0.4
websockets>=12.0
anthropic>=0.39.0
chromadb>=0.4.22
reportlab>=4.0.0
python-multipart>=0.0.6
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
